+++ docs/history.rst
@@ -3,6 +3,7 @@ History
 
 Latest
 ------
+- BUG: Raise OverflowError when nodata data type conversion is unsafe (pull #782)
 
 0.15.5
 ------